First attempt of search caching.
